refactor: :truck: move Modal component to shared folder
